+++ TaskEdicola/Classes/Rivista.cs
@@ -14,7 +14,7 @@ namespace TaskEdicola.Classes
           
            
         }
-        public Rivista(string? titolo, DateTime dataPubblicazione, string? categoria, float prezzo, string codice)
+        public Rivista(string? titolo, DateTime dataPubblicazione, string? categoria, double prezzo, string codice)
         {
             Titolo = titolo;
             DataPubblicazione = dataPubblicazione;
